LPS-144606 Move unit tests as well
