+++ modules/util/source-formatter/source-formatter.properties
@@ -7,7 +7,7 @@
 
     source.formatter.excludes=\
         modules/util/source-formatter/documentation/source-formatter/**,\
-        modules/util/source-formatter/src/test/java/com/liferay/source/formatter/JavaSourceProcessorTest.java
+        modules/util/source-formatter/src/test/java/com/liferay/source/formatter/processors/JavaSourceProcessorTest.java
 
     source.check.JavaExceptionCheck.enabled=false
