+++ modules/util/source-formatter/src/test/java/com/liferay/source/formatter/processors/GradleSourceProcessorTest.java
@@ -12,18 +12,27 @@
  * details.
  */
 
-package com.liferay.source.formatter;
+package com.liferay.source.formatter.processors;
 
 import org.junit.Test;
 
 /**
  * @author Alan Huang
  */
-public class CSSSourceProcessorTest extends BaseSourceProcessorTestCase {
+public class GradleSourceProcessorTest extends BaseSourceProcessorTestCase {
 
 	@Test
-	public void testIncorrectWhitespace() throws Exception {
-		//test("IncorrectWhitespace.testcss");
+	public void testSortMapKeys() throws Exception {
+		test("SortMapKeys1.testgradle");
+		test("SortMapKeys2.testgradle");
+		test("SortMapKeys3.testgradle");
+		test("SortMapKeys4.testgradle");
+		test("SortMapKeys5.testgradle");
+	}
+
+	@Test
+	public void testStylingCheck() throws Exception {
+		test("StylingCheck.testgradle");
 	}
 
 }
