+++ modules/util/source-formatter/src/test/java/com/liferay/source/formatter/processors/HTMLSourceProcessorTest.java
@@ -12,18 +12,23 @@
  * details.
  */
 
-package com.liferay.source.formatter;
+package com.liferay.source.formatter.processors;
 
 import org.junit.Test;
 
 /**
  * @author Alan Huang
  */
-public class PoshiSourceProcessorTest extends BaseSourceProcessorTestCase {
+public class HTMLSourceProcessorTest extends BaseSourceProcessorTestCase {
 
 	@Test
-	public void testIncorrectComments() throws Exception {
-		test("IncorrectComments.testmacro");
+	public void testFormatSelfClosingTags() throws Exception {
+		test("FormatSelfClosingTags.testhtml");
+	}
+
+	@Test
+	public void testIncorrectEmptyLines() throws Exception {
+		test("IncorrectEmptyLines.testhtml");
 	}
 
 }
