+++ modules/util/source-formatter/src/test/java/com/liferay/source/formatter/processors/MarkdownSourceProcessorTest.java
@@ -12,27 +12,18 @@
  * details.
  */
 
-package com.liferay.source.formatter;
+package com.liferay.source.formatter.processors;
 
 import org.junit.Test;
 
 /**
  * @author Alan Huang
  */
-public class GradleSourceProcessorTest extends BaseSourceProcessorTestCase {
+public class MarkdownSourceProcessorTest extends BaseSourceProcessorTestCase {
 
 	@Test
-	public void testSortMapKeys() throws Exception {
-		test("SortMapKeys1.testgradle");
-		test("SortMapKeys2.testgradle");
-		test("SortMapKeys3.testgradle");
-		test("SortMapKeys4.testgradle");
-		test("SortMapKeys5.testgradle");
-	}
-
-	@Test
-	public void testStylingCheck() throws Exception {
-		test("StylingCheck.testgradle");
+	public void testMissingEmptyLines() throws Exception {
+		test("MissingEmptyLines.testmarkdown");
 	}
 
 }
